backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask-cors
PyJWT
werkzeug
python-dotenv
bcrypt
